Clear expo boilerplate and install nativewind
